♻️ HOME: Add <Footer />
